requirements.txt
llama-cpp-python
sentence-transformers
faiss-cpu
numpy
PyPDF2
